+++ setup.py
@@ -52,7 +52,12 @@ setup(
                 'License :: OSI Approved :: BSD License',
                 ],
         entry_points={
-                'napari.plugin': ['zarpaint = zarpaint.plugin'],
-                'console_scripts': ['zarpaint = zarpaint._main:main']
-                },
+             "napari.manifest": [
+                 "zarpaint = zarpaint:napari.yaml",
+             ],
+             'console_scripts': [
+                 'zarpaint = zarpaint._main:main'
+            ]
+         },
+         package_data={"zarpaint": ["napari.yaml"]},
         )
